Enabling .NET 9 tests for SF SDK packages (#124)
This PR enables tests for .NET 9 builds. 

Additionally, we are moving away from the `buils.ps1` script, which uses
`msbuild` under the hood. Instead, we are moving to the `dotnet` tool,
in the aim of streamlining the build process and infrastructure. Tests for multiple target frameworks will be introduced in a future commit.

Internal ADO work item:
https://dev.azure.com/msazure/One/_workitems/edit/29603235